[#1375] Revise public API
move FeaturesResource and Factory to internal package
