+++ bundles/org.eclipse.passage.lic.features.model/META-INF/MANIFEST.MF
@@ -17,6 +17,7 @@ Export-Package: org.eclipse.passage.lic.features.model.api,
  org.eclipse.passage.lic.features.model.impl;x-internal:=true,
  org.eclipse.passage.lic.features.model.meta,
  org.eclipse.passage.lic.features.model.util,
- org.eclipse.passage.lic.internal.features.model;x-friends:="org.eclipse.passage.loc.dashboard.ui"
+ org.eclipse.passage.lic.internal.features.model;x-friends:="org.eclipse.passage.loc.dashboard.ui",
+ org.eclipse.passage.lic.internal.features.model.util;x-internal:=true
 Bundle-ClassPath: .
 Bundle-ActivationPolicy: lazy
