+++ bundles/org.eclipse.passage.lic.features.model/src/org/eclipse/passage/lic/internal/features/model/util/FeaturesResourceFactoryImpl.java
@@ -1,0 +1,26 @@
+/*******************************************************************************
+ * Copyright (c) 2018, 2024 ArSysOp
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * https://www.eclipse.org/legal/epl-2.0/.
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     ArSysOp - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.passage.lic.internal.features.model.util;
+
+import org.eclipse.emf.common.util.URI;
+import org.eclipse.emf.ecore.resource.Resource;
+import org.eclipse.emf.ecore.resource.impl.ResourceFactoryImpl;
+
+public class FeaturesResourceFactoryImpl extends ResourceFactoryImpl {
+
+	@Override
+	public Resource createResource(URI uri) {
+		return new FeaturesResourceImpl(uri);
+	}
+
+}
